review: 1:1 mapping between a logical component’s kind <-> a java package
