+++ statefun-flink/statefun-flink-io-bundle/src/test/resources/kinesis-io-binders/routable-kinesis-ingress-v1.yaml
